Created time Test

+++ src/main/java/org/example/model/Ort.java
@@ -1,11 +1,11 @@
 package org.example.model;
 
 public class Ort {
-    private int plz;
+    private long plz;
     private String name;
     private String region;
 
-    public Ort (int PLZ, String name, String region) {
+    public Ort (long PLZ, String name, String region) {
         this.plz = PLZ;
         this.name = name;
         this.region = region;
